Fixed all warnings

+++ SGP_Honor/Source/Activator.cpp
@@ -54,16 +54,16 @@ void Activator::Update(float elapsedTime)
 	{
 		
 		RECT rSwingRect;
-		rSwingRect.left = m_pPlayer->GetSword()->GetRect().left;
-		rSwingRect.top = m_pPlayer->GetSword()->GetRect().top;
-		rSwingRect.right = m_pPlayer->GetSword()->GetRect().right;
-		rSwingRect.bottom = m_pPlayer->GetSword()->GetRect().bottom;
+		rSwingRect.left = (LONG)m_pPlayer->GetSword()->GetRect().left;
+		rSwingRect.top = (LONG)m_pPlayer->GetSword()->GetRect().top;
+		rSwingRect.right = (LONG)m_pPlayer->GetSword()->GetRect().right;
+		rSwingRect.bottom = (LONG)m_pPlayer->GetSword()->GetRect().bottom;
 
 		RECT rObject;
-		rObject.left = GetRect().left - Camera::GetInstance()->GetCameraPos().x;
-		rObject.top = GetRect().top - Camera::GetInstance()->GetCameraPos().y;
-		rObject.right = GetRect().right - Camera::GetInstance()->GetCameraPos().x;
-		rObject.bottom = GetRect().bottom - Camera::GetInstance()->GetCameraPos().y;
+		rObject.left = (LONG)(GetRect().left - Camera::GetInstance()->GetCameraPos().x);
+		rObject.top = (LONG)(GetRect().top - Camera::GetInstance()->GetCameraPos().y);
+		rObject.right = (LONG)(GetRect().right - Camera::GetInstance()->GetCameraPos().x);
+		rObject.bottom = (LONG)(GetRect().bottom - Camera::GetInstance()->GetCameraPos().y);
 
 		RECT rIntersection = {};
 
@@ -208,16 +208,16 @@ void Activator::HandleCollision(const IEntity* pOther)
 	{
 		
 		RECT rSwingRect;
-		rSwingRect.left =	m_pPlayer->GetSwingRect().left;
-		rSwingRect.top =	m_pPlayer->GetSwingRect().top;
-		rSwingRect.right =	m_pPlayer->GetSwingRect().right;
-		rSwingRect.bottom = m_pPlayer->GetSwingRect().bottom;
+		rSwingRect.left = (LONG)m_pPlayer->GetSwingRect().left;
+		rSwingRect.top = (LONG)m_pPlayer->GetSwingRect().top;
+		rSwingRect.right = (LONG)m_pPlayer->GetSwingRect().right;
+		rSwingRect.bottom = (LONG)m_pPlayer->GetSwingRect().bottom;
 
 		RECT rObject;
-		rObject.left = GetRect().left - Camera::GetInstance()->GetCameraPos().x;
-		rObject.top = GetRect().top - Camera::GetInstance()->GetCameraPos().y;
-		rObject.right = GetRect().right - Camera::GetInstance()->GetCameraPos().x;
-		rObject.bottom = GetRect().bottom - Camera::GetInstance()->GetCameraPos().y;
+		rObject.left = (LONG)(GetRect().left - Camera::GetInstance()->GetCameraPos().x);
+		rObject.top = (LONG)(GetRect().top - Camera::GetInstance()->GetCameraPos().y);
+		rObject.right = (LONG)(GetRect().right - Camera::GetInstance()->GetCameraPos().x);
+		rObject.bottom = (LONG)(GetRect().bottom - Camera::GetInstance()->GetCameraPos().y);
 
 		RECT rIntersection = {};
